Updated, tests in integration component are updated
