--- packages/tests/specs/integration/editor/address/address_create_outofview_spec.json
@@ -1,36 +1,0 @@
-{
-    "layers": [{
-        "id": "linkLayer",
-        "provider": {
-            "id": "linkProvider",
-            "type": "TestProvider",
-            "name": "Links",
-            "level": 15,
-            "class": "NAVLINK"
-        },
-        "min": 14,
-        "max": 20,
-        "clear": {
-            "minLon": 76.5154784445956,
-            "minLat": 12.461322064912764,
-            "maxLon": 76.51976997901943,
-            "maxLat": 12.464464871363163
-        }
-    },{
-        "id": "paLayer",
-        "provider": {
-            "id": "paProvider",
-            "type": "TestProvider",
-            "name": "address",
-            "level": 14
-        },
-        "min": 14,
-        "max": 20,
-        "clear": {
-            "minLon": 76.5154784445956,
-            "minLat": 12.461322064912764,
-            "maxLon": 76.51976997901943,
-            "maxLat": 12.464464871363163
-        }
-    }]
-}
