--- packages/tests/specs/integration/editor/drawingmanager/drawingmanager_create_and_connect_to_original_spec.json
@@ -1,32 +1,0 @@
-{
-    "layers": [{
-        "id": "linkLayer",
-        "provider": {
-            "id": "linkProvider",
-            "type": "TestProvider",
-            "name": "Links",
-            "level": 15,
-            "class": "NAVLINK"
-        },
-        "min": 14,
-        "max": 20,
-        "data": {
-            "remote": [{
-                "id": -188847,
-                "type": "Feature",
-                "properties": {
-                    "featureClass": "NAVLINK",
-                    "type": "residential"
-                },
-                "geometry": {
-                    "type": "LineString",
-                    "coordinates": [
-                        [77.220259106, 13.152202639, 0],
-                        [77.219722664, 13.152202639, 0]
-                    ]
-                }
-            }]
-        },
-        "clear": {"minLon":77.21918622215452,"minLat":13.1495897122955,"maxLon":77.22347775657835,"maxLat":13.152725221581113}
-    }]
-}
