--- packages/tests/specs/integration/editor/drawingmanager/drawingmanager_create_and_split_original_outofview_spec.json
@@ -1,32 +1,0 @@
-{
-    "layers": [{
-        "id": "linkLayer",
-        "provider": {
-            "id": "linkProvider",
-            "type": "TestProvider",
-            "name": "Links",
-            "level": 15,
-            "class": "NAVLINK"
-        },
-        "min": 14,
-        "max": 20,
-        "data": {
-            "remote": [{
-                "id": -188848,
-                "type": "Feature",
-                "properties": {
-                    "featureClass": "NAVLINK",
-                    "type": "residential"
-                },
-                "geometry": {
-                    "type": "LineString",
-                    "coordinates": [
-                        [79.262377116, 13.049935177, 0],
-                        [79.261840675, 13.049935177, 0]
-                    ]
-                }
-            }]
-        },
-        "clear": {"minLon":79.26130423278812,"minLat":13.047322226004937,"maxLon":79.26559576721195,"maxLat":13.050457764052638}
-    }]
-}
